Remove storage and use of seed and use numpy.random.state
We are already passing around the numpy.random.state, so we should do
this to initialize our seed and pass around state instead of seed. All
references of to random number seeds have been removed except for the
one that is stored as a parameter as this seems to hold information
about the number of simualtions.